save point
reworking this method because somewhere the math is getting all sorts of screwed up...I request 12 items from Item#2, my console.writeline() reads that the math was done perfectly, but avallableQnty is showing the wrong value...for example: I requested -12 items, there are 10 in inventory. Console.writeline() (line 36 && 49) show -2 to be the inventory...technically correct we have taken 12 items out of the 10 we had in inventory, however the item.displayItem() shows -22 remaining. SO I am going to 86 this method and completely rewrite it :)