finished invoice
updated remove item to update the total as well, passing the negative of the requested quantity.
worked out printInvoice() as best I could...in hind sight dealing with this as another list was not the best idea but i ran out of time to implement a proper fix. everything else still works as expected.